add docs, move scoring functions to scoring.ts, move experiment naming to utils.ts
